Added 'RunAsPreJob' input and updated tests
