+++ Tasks/AzureKeyVaultV1/task.loc.json
@@ -51,6 +51,17 @@
       "options": {
         "EditableOptions": "True"
       }
+    },
+    {
+      "name": "RunAsPreJob",
+      "type": "boolean",
+      "label": "ms-resource:loc.input.label.RunAsPreJob",
+      "defaultValue": false,
+      "required": true,
+      "helpMarkDown": "ms-resource:loc.input.help.RunAsPreJob",
+      "options": {
+        "EditableOptions": "True"
+      }
     }
   ],
   "dataSourceBindings": [
@@ -64,7 +75,12 @@
   "instanceNameFormat": "ms-resource:loc.instanceNameFormat",
   "prejobexecution": {
     "Node": {
-      "target": "main.js"
+      "target": "runprejob.js"
+    }
+  },
+  "execution": {
+    "Node": {
+      "target": "run.js"
     }
   },
   "messages": {
